requirements.txt
ebooklib
beautifulsoup4
num2words
openai
python-dotenv
tiktoken